Backport: Add fluent-release package for yum/dnf update (#845)
rpm: add fluent-release package for yum/dnf update

In the previous versions, .repo file was assumed to
be created via fluent-package installation script.

For migration from packages.treasuredata.com to fluentd.cdn.cncf.io
in the future, it might be better to setup proper fluent-release
package for it.

Changed in backport

* baseurl was changed to LTS
* fluent-package.repo => fluent-package-lts.repo
* changed section name (fluent-package-lts => fluent-package-lts-v5)
  in fluent-package-lts.repo
* fluent-package-lts-v5 section in fluent-package-lts.repo
  was enabled by default

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>
Co-authored-by: Daijiro Fukuda <fukuda@clear-code.com>
Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>